requirements.txt
pyttsx3~=2.90
pywhatkit~=4.6
wikipedia~=1.4.0
pyjokes~=0.6.0
Eel~=0.12.4
pandas~=1.3.3
numpy~=1.21.2
joblib~=1.0.1
scikit-learn~=0.24.2
pygal~=3.0.0
matplotlib~=3.4.3
plotly~=5.4.0
ipython~=7.30.0
mpld3~=0.5.6
bokeh~=2.4.2
setuptools~=59.5.0
graphviz~=0.19.1
pydotplus~=2.0.2
Pillow~=8.4.0
six~=1.15.0